Moving a lot of things around.
This splits up the 1800 line bzip2.c file into the separate 'reader.c' and 'writer.c' files. This should make it a bit easier to read because 1800 lines isn't very easy on the eyes...
